QC flags to auxiliary; simple QC functions to next_level_qc
